#10 fixed old design choice
at first, the value of a newly made transition was decided by the amount of transitions coming from the same state, which meant it could overwrite other transitions (eg. state q0 goes to q1 via '1', when making a new transition it chose '1' and took away the other transition). this also meant that the inputs could become multidigit, which is unwanted behaviour. this has now be changed to using the first unused value in the alphabet.